refactor taskConsoleTab
